fix: fix linux build

+++ src/gui/errorspopup.cpp
@@ -1,6 +1,6 @@
 /*
  * Infomaniak kDrive - Desktop
- * Copyright (C) 2023-2025 Infomaniak Network SA
+ * Copyright (C) 2023-2026 Infomaniak Network SA
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -79,7 +79,7 @@ ErrorsPopup::ErrorsPopup(const QList<DriveError> &driveErrorList, int genericErr
     // Drive errors
     for (auto const &driveError: driveErrorList) {
         ClickableWidget *driveWidget = new ClickableWidget(this);
-        driveWidget->setProperty(actionTypeProperty.c_str(), driveError.driveDbId);
+        driveWidget->setProperty(actionTypeProperty.c_str(), static_cast<qint64>(driveError.driveDbId));
         mainVBox->addWidget(driveWidget);
 
         QHBoxLayout *driveErrorHBox = new QHBoxLayout();
